(*) #1088 Copied and renamed SuspensionContext to ExtendedSuspensionContext for FastBindingList compability (*) #1088 Moved closer to the implementation of NotifyCollectionChangedEventArgs with NotifyRangedCollectionChangedEventArgs and reduced properties
